requirements.txt
-f /usr/share/pip-wheels
cairocffi==0.8.0
CairoSVG==2.1.3
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0
django-cors-headers==3.1.1
django-filter==2.2.0
django-widget-tweaks==1.4.1
djangorestframework==3.10.3
html5lib==1.0.1
image==1.5.17
Markdown==3.1.1
pdfrw==0.4
Pillow==5.0.0
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2017.3
reportlab==3.4.0
six==1.11.0
tinycss2==0.6.1
WeasyPrint==0.42.2
webencodings==0.5.1
